% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemblyai.R
\name{assemblyai_format_text_from_utterances}
\alias{assemblyai_format_text_from_utterances}
\title{Format text from utterances}
\usage{
assemblyai_format_text_from_utterances(
  utterances,
  include_timestamps,
  timestamp_format
)
}
\description{
Format text from utterances
}
\keyword{internal}
