% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemblyai.R
\name{assemblyai_is_authenticated}
\alias{assemblyai_is_authenticated}
\title{Check if authenticated to AssemblyAI API}
\usage{
assemblyai_is_authenticated()
}
\value{
Logical value indicating whether an API key is stored.
}
\description{
Check whether an API key has been set for the current session.
}
\examples{
\dontrun{
assemblyai_is_authenticated()
}

}
