% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemblyai.R
\docType{data}
\name{.assemblyai_env}
\alias{.assemblyai_env}
\title{Package environment to store API credentials}
\format{
An object of class \code{environment} of length 0.
}
\usage{
.assemblyai_env
}
\description{
Package environment to store API credentials
}
\keyword{internal}
