% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mistral.R
\docType{data}
\name{.mistral_env}
\alias{.mistral_env}
\title{Mistral package environment}
\format{
An object of class \code{environment} of length 0.
}
\usage{
.mistral_env
}
\description{
Internal environment for storing the API key.
}
\keyword{internal}
