% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcript_formats.R
\name{elevenlabs_format_text_with_metadata}
\alias{elevenlabs_format_text_with_metadata}
\title{Format text with metadata (timestamps and speakers)}
\usage{
elevenlabs_format_text_with_metadata(
  words,
  include_timestamps,
  include_speakers,
  timestamp_format
)
}
\description{
Format text with metadata (timestamps and speakers)
}
\keyword{internal}
