% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate.R
\name{elevenlabs_is_authenticated}
\alias{elevenlabs_is_authenticated}
\title{Check if authenticated to ElevenLabs API}
\usage{
elevenlabs_is_authenticated()
}
\value{
Logical value indicating whether an API key is stored.
}
\description{
Check whether an API key has been set for the current session.
}
\examples{
\dontrun{
elevenlabs_is_authenticated()
}

}
