% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mistral.R
\name{mistral_check_timestamps}
\alias{mistral_check_timestamps}
\title{Validate requested timestamp granularities}
\usage{
mistral_check_timestamps(timestamps)
}
\arguments{
\item{timestamps}{Character vector or NULL.}
}
\value{
The validated vector, or NULL.
}
\description{
Validate requested timestamp granularities
}
\keyword{internal}
