% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mistral.R
\name{mistral_format_time_timestamp}
\alias{mistral_format_time_timestamp}
\title{Format timestamp for readable time format (Mistral version)}
\usage{
mistral_format_time_timestamp(seconds)
}
\description{
Format timestamp for readable time format (Mistral version)
}
\keyword{internal}
