% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mistral.R
\name{mistral_is_authenticated}
\alias{mistral_is_authenticated}
\title{Check if authenticated to Mistral API}
\usage{
mistral_is_authenticated()
}
\value{
Logical value indicating whether an API key is stored.
}
\description{
Check whether an API key has been set for the current session.
}
\examples{
\dontrun{
mistral_is_authenticated()
}

}
